Fix slow matrix assembly with triplet lists
